requirements.txt
tweepy
pandas
nltk
PySastrawi
kafka-python